% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na.R
\name{check_na}
\alias{check_na}
\title{Return missingness info for a variable}
\usage{
check_na(variable)
}
\arguments{
\item{variable}{Numeric, complex, or logical vectors.}
}
\value{
A matrix including the number of total, missing, and non-missing
  observations
}
\description{
Take variable as input and compute the number of total observations, number
of missing observations and number of non-missing observations.
}
\examples{
check_na(c(NA, NA, 3, "cbds!", 4.5))
check_na(iris$Sepal.Width)
}
